+++ config/projects/chef-server.rb
@@ -30,14 +30,15 @@ override :rebar, version: "2.0.0"
 override :berkshelf2, version: "2.0.18"
 override :rabbitmq, version: "3.3.4"
 override :erlang, version: "R16B03-1"
-override :'omnibus-ctl', version: "0.3.1"
-override :'chef-gem', version: "11.18.0"
+override :ruby, version: "2.1.4"
+override :'omnibus-ctl', version: "0.3.2"
+override :'chef-gem', version: "12.0.3"
 
 # creates required build directories
 dependency "preparation"
 
 # global
-dependency "chef-gem" # for embedded chef-solo
+dependency "chef-gem" # for embedded chef-client -z runs
 dependency "private-chef-scripts" # assorted scripts used by installed instance
 dependency "private-chef-ctl" # additional project-specific private-chef-ctl subcommands
 dependency "ctl-man" # install man page