+++ config/software/knife-ec-backup-gem.rb
@@ -15,7 +15,7 @@
 #
 
 name "knife-ec-backup"
-default_version "2.0.0.beta.2"
+default_version "2.0.1"
 
 dependency "pg-gem"
 dependency "sequel-gem"